adding/validating sat db utils
