+++ setup.py
@@ -50,6 +50,7 @@ setuptools.setup(
             'swell_task = swell.tasks.base.task_base:main',
             'swell_create_experiment = swell.deployment.bin.swell_create_experiment:main',
             'swell_launch_experiment = swell.deployment.bin.swell_launch_experiment:main',
+            'swell_sat_db_processing = swell.deployment.bin.swell_sat_db_processing:main',
         ],
     },
     )
